requirements.txt
flask
openai-whisper
ffmpeg-python